fix: Change kindgarden_id type from unsignedBigInteger to unsignedInteger in contracts and requisites tables

+++ database/migrations/2026_02_23_000002_create_kindgarden_requisites_table.php
@@ -10,7 +10,7 @@ class CreateKindgardenRequisitesTable extends Migration
     {
         Schema::create('kindgarden_requisites', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('kindgarden_id');
+            $table->unsignedInteger('kindgarden_id');
             $table->string('director_name')->nullable();
             $table->string('director_phone')->nullable();
             $table->string('reception_phone')->nullable();
